+++ src/Standards/Squiz/Sniffs/Commenting/BlockCommentSniff.php
@@ -256,7 +256,7 @@ class BlockCommentSniff implements Sniff
                         && $tokens[$commentLines[1]]['orig_content'][0] === "\t"
                     ) {
                         // Line is indented using tabs.
-                        $padding  = str_repeat("\t", floor($expected / $this->tabWidth));
+                        $padding  = str_repeat("\t", (int) floor($expected / $this->tabWidth));
                         $padding .= str_repeat(' ', ($expected % $this->tabWidth));
                     } else {
                         $padding = str_repeat(' ', $expected);
@@ -316,7 +316,7 @@ class BlockCommentSniff implements Sniff
                         && $tokens[$line]['orig_content'][0] === "\t"
                     ) {
                         // Line is indented using tabs.
-                        $padding  = str_repeat("\t", floor($expected / $this->tabWidth));
+                        $padding  = str_repeat("\t", (int) floor($expected / $this->tabWidth));
                         $padding .= str_repeat(' ', ($expected % $this->tabWidth));
                     } else {
                         $padding = str_repeat(' ', $expected);
@@ -360,7 +360,7 @@ class BlockCommentSniff implements Sniff
                         && $tokens[$line]['orig_content'][0] === "\t"
                     ) {
                         // Line is indented using tabs.
-                        $padding  = str_repeat("\t", floor($expected / $this->tabWidth));
+                        $padding  = str_repeat("\t", (int) floor($expected / $this->tabWidth));
                         $padding .= str_repeat(' ', ($expected % $this->tabWidth));
                     } else {
                         $padding = str_repeat(' ', $expected);
